[Exceptions]: Se crearon excepciones para la gran mayoría de metodos y se dejo de usar entityManager.find() para realizar consultas

+++ ambiente/semillero-hbt/semillero-padre/semillero-servicios/src/main/java/com/hbt/semillero/interfaces/ICalcularPrecioIVA.java
@@ -17,7 +17,7 @@ public interface ICalcularPrecioIVA {
 	 * @param precio precio base del comic
 	 * @return
 	 */
-	public double CalcularPrecioTotal(double iva, BigDecimal precio);
+	public BigDecimal CalcularPrecioTotal(double iva, BigDecimal precio);
 	
 
 	/**
